refactor(llm): move concurrency control to llm module
- Move ConcurrencyConfig and ConcurrencyController from root throttle
  module to llm::throttle submodule
- Remove the standalone throttle module and integrate its functionality
  directly into the llm module
- Update imports across related files to use the new path
  crate::llm::throttle instead of crate::throttle
- Add comprehensive documentation and tests for the throttle
  functionality within the llm module
