+++ rust/src/index/config.rs
@@ -12,7 +12,7 @@
 use super::summary::SummaryStrategy;
 use crate::config::IndexerConfig;
 use crate::document::{DocumentTree, ReasoningIndexConfig};
-use crate::throttle::ConcurrencyConfig;
+use crate::llm::throttle::ConcurrencyConfig;
 use crate::utils::fingerprint::{Fingerprint, Fingerprinter};
 
 use std::path::PathBuf;
