+++ rust/src/llm/client.rs
@@ -12,7 +12,7 @@ use super::config::LlmConfig;
 use super::error::{LlmError, LlmResult};
 use super::executor::LlmExecutor;
 use super::fallback::FallbackChain;
-use crate::throttle::ConcurrencyController;
+use super::throttle::ConcurrencyController;
 
 /// Unified LLM client.
 ///
@@ -355,7 +355,7 @@ mod tests {
 
     #[test]
     fn test_client_with_concurrency() {
-        use crate::throttle::ConcurrencyConfig;
+        use super::throttle::ConcurrencyConfig;
 
         let controller = ConcurrencyController::new(ConcurrencyConfig::conservative());
         let client = LlmClient::for_model("gpt-4o-mini").with_concurrency(controller);
