+++ rust/src/llm/mod.rs
@@ -34,8 +34,10 @@ mod error;
 mod executor;
 mod fallback;
 mod pool;
+pub(crate) mod throttle;
 
 pub use client::LlmClient;
 pub use error::LlmResult;
 pub use executor::LlmExecutor;
 pub use pool::LlmPool;
+pub(crate) use throttle::{ConcurrencyConfig, ConcurrencyController};
